Create the AuthModal component.
Create a state that determines whether the modal is shown (in Home.js).
If the create account button is clicked, then setShowModal to be true.
Pass in setShowModal as a prop for the AuthModal component.
In AuthModal, if the X button is clicked, then the auth modal will
disappear (set state to be false).